design fised

+++ src/Layouts/MyOrders/MyOrders.jsx
@@ -2,6 +2,7 @@ import axios from 'axios';
 import { useEffect, useState } from 'react';
 import toast from 'react-hot-toast';
 import { useLoaderData } from 'react-router-dom';
+import Swal from 'sweetalert2';
 
 const MyOrders = () => {
     const data = useLoaderData()
@@ -21,24 +22,43 @@ const MyOrders = () => {
     const [ramaing,setRemaining] = useState(data)
 
     const handleDelete =(_id)=>{
-        axios.delete(`https://restaurant-management-server-site.vercel.app/purchaseFood/${_id}`)
-        .then(res=>{
-           
-            if(res.data.deletedCount>0){
-                toast.success("Deleted Successfully")
-
-                const remain = ramaing.filter(item=> item._id!== _id)
-                setRemaining(remain)
-            }
+        Swal.fire({
+            title: "Are you sure?",
+            text: "You won't be able to revert this!",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#3085d6",
+            cancelButtonColor: "#d33",
+            confirmButtonText: "Yes, delete it!"
+          }).then((result) => {
             
-        })
+              axios.delete(`https://restaurant-management-server-site.vercel.app/purchaseFood/${_id}`)
+              .then(res=>{
+                 
+                  if(res.data.deletedCount>0){
+                    if (result.isConfirmed) {
+                        Swal.fire({
+                          title: "Deleted!",
+                          text: "Your file has been deleted.",
+                          icon: "success"
+                        });
+                      }
+      
+                      const remain = ramaing.filter(item=> item._id!== _id)
+                      setRemaining(remain)
+                  }
+                  
+              })
+          });
     }
     
 
 
     return (
         <div>
-            <div className="overflow-x-auto">
+
+            <p className='text-3xl font-semibold text-center py-10'>My Orders Foods</p>
+            <div className="overflow-x-auto w-[80%] m-auto pb-20">
                 <table className="table">
                     {/* head */}
                     <thead>
